feat(*): refactor views
