+++ src/assignment/views/AssignmentDetail.scss
