+++ src/assignment/views/AssignmentEdit.scss
