+++ src/assignment/views/AssignmentEdit.tsx
@@ -41,38 +41,38 @@ import {
 import { Avo } from '@viaa/avo2-types';
 import { AssignmentContent } from '@viaa/avo2-types/types/assignment/types';
 
-import { getProfileId, getProfileName } from '../../../authentication/helpers/get-profile-info';
-import { selectLogin } from '../../../authentication/store/selectors';
-import { LoginResponse } from '../../../authentication/store/types';
-import { CollectionService } from '../../../collection/collection.service';
-import { DutchContentType, toEnglishContentType } from '../../../collection/collection.types';
-import { RouteParts } from '../../../constants';
+import { getProfileId, getProfileName } from '../../authentication/helpers/get-profile-info';
+import { selectLogin } from '../../authentication/store/selectors';
+import { LoginResponse } from '../../authentication/store/types';
+import { CollectionService } from '../../collection/collection.service';
+import { DutchContentType, toEnglishContentType } from '../../collection/collection.types';
+import { RouteParts } from '../../constants';
 import {
 	DeleteObjectModal,
 	InputModal,
 	LoadingErrorLoadedComponent,
-} from '../../../shared/components';
-import { renderDropdownButton } from '../../../shared/components/CheckboxDropdownModal/CheckboxDropdownModal';
-import { copyToClipboard } from '../../../shared/helpers';
-import { dataService } from '../../../shared/services/data-service';
-import { EventObjectType, trackEvents } from '../../../shared/services/event-logging-service';
-import toastService, { TOAST_TYPE } from '../../../shared/services/toast-service';
+} from '../../shared/components';
+import { renderDropdownButton } from '../../shared/components/CheckboxDropdownModal/CheckboxDropdownModal';
+import { copyToClipboard } from '../../shared/helpers';
+import { dataService } from '../../shared/services/data-service';
+import { EventObjectType, trackEvents } from '../../shared/services/event-logging-service';
+import toastService, { TOAST_TYPE } from '../../shared/services/toast-service';
 
-import { deleteAssignment, insertAssignment, updateAssignment } from '../../assignment.services';
-import { AssignmentLayout } from '../../assignment.types';
+import { deleteAssignment, insertAssignment, updateAssignment } from '../assignment.services';
+import { AssignmentLayout } from '../assignment.types';
 
 import {
 	GET_COLLECTION_BY_ID,
 	INSERT_COLLECTION,
 	INSERT_COLLECTION_FRAGMENTS,
-} from '../../../collection/collection.gql';
-import { GET_ITEM_BY_ID } from '../../../item/item.gql';
+} from '../../collection/collection.gql';
+import { GET_ITEM_BY_ID } from '../../item/item.gql';
 import {
 	DELETE_ASSIGNMENT,
 	GET_ASSIGNMENT_BY_ID,
 	INSERT_ASSIGNMENT,
 	UPDATE_ASSIGNMENT,
-} from '../../assignment.gql';
+} from '../assignment.gql';
 
 import './AssignmentEdit.scss';
