+++ src/authentication/views/Logout.tsx
@@ -2,9 +2,9 @@ import { FunctionComponent } from 'react';
 
 import { RouteComponentProps } from 'react-router';
 
-import { RouteParts } from '../../../constants';
+import { RouteParts } from '../../constants';
 
-import { redirectToLogoutPage } from '../../helpers/redirect-to-idp';
+import { redirectToLogoutPage } from '../helpers/redirect-to-idp';
 
 export interface LogoutProps extends RouteComponentProps {}
