+++ src/authentication/views/Register.scss
