+++ src/authentication/views/Register.tsx
