+++ src/authentication/views/RegisterOrLogin.scss
@@ -1,5 +1,5 @@
 @mixin background-smartschool {
-	background: url("/images/smartschool-login.png");
+	background: url('/images/smartschool-login.png');
 	background-size: cover;
 }
