+++ src/authentication/views/index.ts
@@ -1,6 +1,6 @@
-import Login from './Login/Login';
-import Logout from './Logout/Logout';
-import Register from './Register/Register';
-import RegisterOrLogin from './RegisterOrLogin/RegisterOrLogin';
+import Login from './Login';
+import Logout from './Logout';
+import Register from './Register';
+import RegisterOrLogin from './RegisterOrLogin';
 
 export { Login, Logout, Register, RegisterOrLogin };
