+++ src/collection/views/CollectionDetail.scss
