+++ src/collection/views/CollectionEditMetaData.tsx
@@ -14,11 +14,11 @@ import {
 import { TagInfo } from '@viaa/avo2-components/dist/components/TagsInput/TagsInput';
 import { Avo } from '@viaa/avo2-types';
 
-import { DataQueryComponent } from '../../../shared/components';
+import { DataQueryComponent } from '../../shared/components';
 
-import { GET_CLASSIFICATIONS_AND_SUBJECTS } from '../../collection.gql';
-import { getValidationFeedbackForShortDescription } from '../../collection.helpers';
-import { CollectionStillsModal } from '../../components';
+import { GET_CLASSIFICATIONS_AND_SUBJECTS } from '../collection.gql';
+import { getValidationFeedbackForShortDescription } from '../collection.helpers';
+import { CollectionStillsModal } from '../components';
 
 interface CollectionEditMetaDataProps {
 	collection: Avo.Collection.Collection;
