+++ src/collection/views/CollectionOverview.scss
