+++ src/collection/views/CollectionOverview.tsx
@@ -21,17 +21,17 @@ import {
 import { Avo } from '@viaa/avo2-types';
 import { compact } from 'lodash-es';
 
-import { getProfileId } from '../../../authentication/helpers/get-profile-info';
-import { RouteParts } from '../../../constants';
-import { ErrorView } from '../../../error/views';
-import { DataQueryComponent, DeleteObjectModal } from '../../../shared/components';
-import { formatDate, formatTimestamp, fromNow, getAvatarProps } from '../../../shared/helpers';
-import { ApolloCacheManager } from '../../../shared/services/data-service';
-import toastService, { TOAST_TYPE } from '../../../shared/services/toast-service';
-import { IconName } from '../../../shared/types/types';
-import { ITEMS_PER_PAGE } from '../../../workspace/workspace.const';
+import { getProfileId } from '../../authentication/helpers/get-profile-info';
+import { RouteParts } from '../../constants';
+import { ErrorView } from '../../error/views';
+import { DataQueryComponent, DeleteObjectModal } from '../../shared/components';
+import { formatDate, formatTimestamp, fromNow, getAvatarProps } from '../../shared/helpers';
+import { ApolloCacheManager } from '../../shared/services/data-service';
+import toastService, { TOAST_TYPE } from '../../shared/services/toast-service';
+import { IconName } from '../../shared/types/types';
+import { ITEMS_PER_PAGE } from '../../workspace/workspace.const';
 
-import { DELETE_COLLECTION, GET_COLLECTIONS_BY_OWNER } from '../../collection.gql';
+import { DELETE_COLLECTION, GET_COLLECTIONS_BY_OWNER } from '../collection.gql';
 
 import './CollectionOverview.scss';
