+++ src/collection/views/index.ts
@@ -1,8 +1,8 @@
-import CollectionDetail from './CollectionDetail/CollectionDetail';
-import CollectionEdit from './CollectionEdit/CollectionEdit';
-import CollectionEditContent from './CollectionEdit/CollectionEditContent';
-import CollectionEditMetaData from './CollectionEdit/CollectionEditMetaData';
-import CollectionOverview from './CollectionOverview/CollectionOverview';
+import CollectionDetail from './CollectionDetail';
+import CollectionEdit from './CollectionEdit';
+import CollectionEditContent from './CollectionEditContent';
+import CollectionEditMetaData from './CollectionEditMetaData';
+import CollectionOverview from './CollectionOverview';
 
 export {
 	CollectionEditContent,
