+++ src/discover/views/Discover.tsx
