+++ src/home/views/index.ts
@@ -1,3 +1,3 @@
-import Home from './Home/Home';
+import Home from './Home';
 
 export { Home };
